[ci skip] use dbaccessors
